move scan helpers, reorganize, add comments
